added test
added a test, I am trying to use an AI for this so I will see if this is a good way of handling this. the processing functionality seems easier to work with, however I am having issues with the actual testing of api functions as I have only tested if the code itself is funcitonal before.