Users & tests
User validation is working and field tested. Currently you only need to be a user to gain access to everything, although in the future I might consider having rules about which modules you are allowed or not allowed to see, especially considering location data is involved.

Testing has been reworked, for now I am just creating some python scripts to handle this because they seemed like the easiest thing to work with at the time.

added the docker compose build chain to the project, hopefully that works.